refactor: extract translation-validation command behavior into a shared trait
